Server/nasahack/requirements.txt
pandas
numpy
dj_database_url
django_heroku